+++ jakub/simtel_baseline.py
@@ -1,31 +1,6 @@
 import numpy as np
 
 
-def baseline_data(event_stream, n_bins0=5, n_bins1=10):
-    for event in event_stream:
-
-        for telescope_id in event.r0.tels_with_data:
-            n_pixels = event.inst.num_pixels[telescope_id]
-            r0_camera = event.r0.tel[telescope_id]
-
-            adc_samples = r0_camera.adc_samples
-
-            # Selection of only n_bins0 first samples and
-            # n_bins1 last samples from given 50 samples
-            adc_samples_first = adc_samples[:, 0:n_bins0 - 1]
-            adc_samples_last = adc_samples[:, -n_bins1:]
-            adc_samples = np.concatenate((adc_samples_first,
-                                          adc_samples_last), axis=1)
-
-            baseline = np.mean(adc_samples, axis=-1)
-            std = np.std(adc_samples, axis=-1)
-
-            r0_camera.baseline = baseline
-            r0_camera.standard_deviation = std
-
-        yield event
-
-
 def baseline_simtel(event_stream):
     for event in event_stream:
